Relocate client netty classes.
