+++ core/client/fs/src/main/java/alluxio/client/block/stream/LocalFilePacketReader.java
@@ -15,8 +15,8 @@ import alluxio.Configuration;
 import alluxio.PropertyKey;
 import alluxio.client.file.FileSystemContext;
 import alluxio.client.file.options.InStreamOptions;
-import alluxio.client.netty.NettyRPC;
-import alluxio.client.netty.NettyRPCContext;
+import alluxio.network.netty.NettyRPC;
+import alluxio.network.netty.NettyRPCContext;
 import alluxio.network.protocol.databuffer.DataBuffer;
 import alluxio.network.protocol.databuffer.DataByteBuffer;
 import alluxio.proto.dataserver.Protocol;
