+++ core/client/fs/src/test/java/alluxio/client/block/AlluxioBlockStoreTest.java
@@ -24,8 +24,8 @@ import alluxio.client.file.options.InStreamOptions;
 import alluxio.client.file.options.OpenFileOptions;
 import alluxio.client.file.options.OutStreamOptions;
 import alluxio.client.file.policy.FileWriteLocationPolicy;
-import alluxio.client.netty.NettyRPC;
-import alluxio.client.netty.NettyRPCContext;
+import alluxio.network.netty.NettyRPC;
+import alluxio.network.netty.NettyRPCContext;
 import alluxio.exception.ExceptionMessage;
 import alluxio.exception.PreconditionMessage;
 import alluxio.exception.status.UnavailableException;
