+++ core/common/src/main/java/alluxio/network/netty/IdleWriteHandler.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.client.netty;
+package alluxio.network.netty;
 
 import alluxio.network.protocol.RPCProtoMessage;
 import alluxio.proto.dataserver.Protocol;
