+++ core/common/src/main/java/alluxio/network/netty/NettyClient.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.client.netty;
+package alluxio.network.netty;
 
 import alluxio.Configuration;
 import alluxio.PropertyKey;
